+++ lib/reactive_forms.dart
@@ -1,18 +1,29 @@
+// Copyright 2020 Joan Pablo Jiménez Milian. All rights reserved.
+// Use of this source code is governed by the MIT license that can be
+// found in the LICENSE file.
+
 library reactive_forms;
 
+export 'exceptions/form_array_invalid_index_exception.dart';
+export 'exceptions/form_control_invalid_name_exception.dart';
+export 'exceptions/form_control_not_found_exception.dart';
 export 'models/abstract_control.dart';
+export 'models/control_status.dart';
 export 'models/form_control.dart';
 export 'models/form_group.dart';
 export 'validators/validation_message.dart';
 export 'validators/validator.dart';
 export 'validators/validators.dart';
 export 'widgets/reactive_checkbox.dart';
+export 'widgets/reactive_checkbox_list_tile.dart';
 export 'widgets/reactive_dropdown_field.dart';
 export 'widgets/reactive_form.dart';
+export 'widgets/reactive_form_array.dart';
 export 'widgets/reactive_form_consumer.dart';
 export 'widgets/reactive_form_field.dart';
 export 'widgets/reactive_radio.dart';
 export 'widgets/reactive_slider.dart';
+export 'widgets/reactive_status_listenable_builder.dart';
 export 'widgets/reactive_switch.dart';
 export 'widgets/reactive_text_field.dart';
 export 'widgets/reactive_value_listenable_builder.dart';