+++ lib/widgets/reactive_status_listenable_builder.dart
@@ -1,0 +1,32 @@
+import 'package:flutter/material.dart';
+import 'package:reactive_forms/models/form_control_collection.dart';
+import 'package:reactive_forms/reactive_forms.dart';
+
+typedef ReactiveStatusListenableWidgetBuilder = Widget Function(
+    BuildContext context, AbstractControl control, Widget child);
+
+class ReactiveStatusListenableBuilder extends StatelessWidget {
+  final String formControlName;
+  final Widget child;
+  final ReactiveStatusListenableWidgetBuilder builder;
+
+  const ReactiveStatusListenableBuilder({
+    Key key,
+    @required this.formControlName,
+    @required this.builder,
+    this.child,
+  })  : assert(formControlName != null),
+        assert(builder != null);
+
+  @override
+  Widget build(BuildContext context) {
+    final form =
+        ReactiveForm.of(context, listen: false) as FormControlCollection;
+    final formControl = form.formControl(this.formControlName);
+    return ValueListenableBuilder<ControlStatus>(
+      valueListenable: formControl.onStatusChanged,
+      builder: (context, status, child) =>
+          this.builder(context, formControl, child),
+    );
+  }
+}