+++ test/validators_test.dart
@@ -62,4 +62,32 @@ void main() {
       );
     });
   });
+
+  group('Pattern Validator Tests', () {
+    test('FormControl invalid if value not matched pattern', () {
+      const AmericanExpressPattern = r'^3[47][0-9]{13}$';
+
+      final cardNumber = FormControl(
+        validators: [Validators.pattern(AmericanExpressPattern)],
+      );
+
+      cardNumber.value = '395465465421'; // not a valid number
+
+      expect(cardNumber.valid, false);
+      expect(cardNumber.errors.containsKey(ValidationMessage.pattern), true);
+    });
+
+    test('FormControl valid if value matched pattern', () {
+      const AmericanExpressPattern = r'^3[47][0-9]{13}$';
+
+      final cardNumber = FormControl(
+        validators: [Validators.pattern(AmericanExpressPattern)],
+      );
+
+      cardNumber.value = '342654321654213';
+
+      expect(cardNumber.valid, true);
+      expect(cardNumber.errors.keys.isEmpty, true);
+    });
+  });
 }